+++ src/solvers/de/de_solver_cpu_impl.hpp
@@ -7,7 +7,7 @@ template <typename TFloat>
 de_solver_cpu<TFloat>::de_solver_cpu(population_set_cpu<TFloat>* population,
                                      evaluator_cpu<TFloat>* evaluator,
                                      prngenerator_cpu<TFloat>* prn_generator,
-                                     uint32_t generation_target,
+                                     uint64_t generation_target,
                                      TFloat* upper_bounds, TFloat* lower_bounds)
 
   : evolutionary_solver_cpu<TFloat>(population, evaluator, prn_generator,
