+++ src/solvers/evolutionary_solver_cuda.hpp
@@ -23,7 +23,7 @@ struct evolutionary_solver_cuda : evolutionary_solver<TFloat>
   evolutionary_solver_cuda(population_set_cuda<TFloat>* population,
                            evaluator_cuda<TFloat>* evaluator,
                            prngenerator_cuda<TFloat>* prn_generator,
-                           uint32_t generation_target, TFloat* upper_bounds,
+                           uint64_t generation_target, TFloat* upper_bounds,
                            TFloat* lower_bounds);
 
   /// Default destructor
