+++ src/solvers/ga/ga_solver_cuda.hpp
@@ -22,7 +22,7 @@ struct ga_solver_cuda : evolutionary_solver_cuda<TFloat>
   ga_solver_cuda(population_set_cuda<TFloat>* population,
                  evaluator_cuda<TFloat>* evaluator,
                  prngenerator_cuda<TFloat>* prn_generator,
-                 uint32_t generation_target, TFloat* upper_bounds,
+                 uint64_t generation_target, TFloat* upper_bounds,
                  TFloat* lower_bounds);
 
   virtual ~ga_solver_cuda();
