+++ src/solvers/pso/pso_solver_cpu.hpp
@@ -23,7 +23,7 @@ struct pso_solver_cpu : evolutionary_solver_cpu<TFloat>
   pso_solver_cpu(population_set_cpu<TFloat>* population,
                  evaluator_cpu<TFloat>* evaluator,
                  prngenerator_cpu<TFloat>* prn_generator,
-                 uint32_t generation_target, TFloat* upper_bounds,
+                 uint64_t generation_target, TFloat* upper_bounds,
                  TFloat* lower_bounds);
 
   virtual ~pso_solver_cpu();
